+++ src/utils.c
@@ -6,7 +6,7 @@
 /*   By: tlouro-c <tlouro-c@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/19 22:45:05 by tlouro-c          #+#    #+#             */
-/*   Updated: 2024/01/06 14:23:24 by tlouro-c         ###   ########.fr       */
+/*   Updated: 2024/01/06 16:05:52 by tlouro-c         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -56,22 +56,6 @@ int	run_builtin(t_cmd *cmd, t_enviroment *enviroment)
 	else
 		cmd_exit(cmd->args, enviroment);
 	return (status);
-}
-
-int	ft_strcmp_heredoc(const char *s1, const char *s2)
-{
-	int	i;
-
-	i = 0;
-	if (s1[i] == '\n' && s2[i] == '\n')
-		return (0);
-	while (s1[i] == s2[i] && s1[i] != '\0' )
-	{
-		if (s1[i + 1] == '\n')
-			break ;
-		i++;
-	}
-	return (s1[i] - s2[i]);
 }
 
 int	ft_key_only_snake(char *s)
